Moved island setup param related code into cc/setup folder
