+++ monkey/monkey_island/setup/setup_param_factory.py
@@ -1,9 +1,8 @@
 import os
 
-from setup_params import SetupParams
-
 from monkey_island import config_file_parser
 from monkey_island.cc.arg_parser import IslandArgs
+from monkey_island.setup.setup_params import SetupParams
 
 
 class SetupParamFactory:
